Fix gallery/edit transitions, implement delete
Long press a fractal in the gallery to show delete button, click
to delete.

Refresh gallery list on delete or on returning to gallery from edit
activity (in case user saved a new fractal).

Initialize FractalStateManager with an empty state, not a Sierpinski
pyramid.

Fix crash on clicking new fractal button from gallery, show a toast
instructing the user to click the + button to add transforms.

Fixes #8